support labels in assembler

+++ src/instructions.h
@@ -229,7 +229,7 @@ class JumpIfFalseInstruction: public JumpInstruction {
 class CallInstruction: public Instruction {
     public:
     CallInstruction();
-    CallInstruction(const uint64_t& address, const uint8_t& param_count);
+    CallInstruction(const uint64_t& address, const uint64_t& param_count);
     void read(const uint8_t* buffer, uint64_t* index);
     void write(std::vector<uint8_t>& buffer);
     void execute(Vm& vm) const;
@@ -238,7 +238,7 @@ class CallInstruction: public Instruction {
 
     private:
     uint64_t address;
-    uint8_t param_count;
+    uint64_t param_count;
 };
 
 class RetInstruction: public Instruction {
